Move validation of correlation between license and author to article-api
