+++ draft-api/src/main/scala/no/ndla/draftapi/validation/ContentValidator.scala
@@ -175,28 +175,18 @@ trait ContentValidator {
 
     private def validateCopyright(copyright: Copyright): Seq[ValidationMessage] = {
       val licenseMessage = copyright.license.map(validateLicense).toSeq.flatten
-      val allAuthors = copyright.creators ++ copyright.processors ++ copyright.rightsholders
-      val licenseCorrelationMessage = validateAuthorLicenseCorrelation(copyright.license, allAuthors)
       val contributorsMessages = copyright.creators.flatMap(validateAuthor) ++ copyright.processors.flatMap(
         validateAuthor) ++ copyright.rightsholders.flatMap(validateAuthor)
       val originMessage =
         copyright.origin.map(origin => NoHtmlValidator.validate("copyright.origin", origin)).toSeq.flatten
 
-      licenseMessage ++ licenseCorrelationMessage ++ contributorsMessages ++ originMessage
+      licenseMessage ++ contributorsMessages ++ originMessage
     }
 
     private def validateLicense(license: String): Seq[ValidationMessage] = {
       getLicense(license) match {
         case None => Seq(ValidationMessage("license.license", s"$license is not a valid license"))
         case _    => Seq()
-      }
-    }
-
-    private def validateAuthorLicenseCorrelation(license: Option[String], authors: Seq[Author]) = {
-      val errorMessage = (lic: String) => ValidationMessage("license.license", s"At least one copyright holder is required when license is $lic")
-      license match {
-        case None => Seq()
-        case Some(lic) => if(lic == "N/A" || authors.nonEmpty) Seq() else Seq(errorMessage(lic))
       }
     }
